requirements.txt
black 
chumpy==0.70 
easydict==1.10 
ftfy==6.1.1 
glfw==2.5.9 
imageio-ffmpeg 
imageio==2.26.1 
iopath==0.1.10 
kaleido==0.2.1 
loguru==0.6.0 
matplotlib==3.7.1 
ninja==1.11.1 
omegaconf==2.3.0 
opencv-contrib-python-headless==4.8.0.74 
pandas==1.5.3 
Pillow==10.0.0 
plyfile==0.8.1 
PyOpenGL_accelerate==3.1.6 
PyOpenGL==3.1.0 
pyrender==0.1.45 
regex==2023.3.23 
scipy==1.8.1 
shapely==2.0.1 
SharedArray==3.2.2 
tensorboardX==2.6 
torchgeometry==0.1.2 
tqdm==4.65.0 
trimesh[easy]
usort 
wandb==0.15.8 
protobuf==3.19.6
tensorflow==2.11.1
accelerate==0.21.0 
datasets==2.14.2 
safetensors==0.3.1 
transformers==4.31.0 
xformers==0.0.20 
diffusers[torch]==0.21.0
git+https://github.com/NVlabs/nvdiffrast.git
